### 2023-07-06 #5622 @Midnighter
Hmm, not sure about the error. The paper compiles fine using the Docker image. It looks like something in the retrieval process is failing.
