### 2023-07-07 #5622 @xuanxu
There is a problem with the format of the date in the paper.md metadata, merging [this PR](https://github.com/taxprofiler/taxpasta/pull/120) should fix it
