### 2023-07-07 #5622 @Kevin-Mattheus-Moerman
The work submitted here has been reviewed/accepted by `pyOpenSci`, see here:  https://github.com/pyOpenSci/software-submission/issues/84
