### 2023-07-07 #5627 @editorialbot
[REVIEW]: TAXPASTA: TAXonomic Profile Aggregation and STAndardisation
**Submitting author:** @Midnighter (<a href="http://orcid.org/0000-0003-2406-1978">Moritz Emanuel Beber</a>)
**Repository:** https://github.com/taxprofiler/taxpasta
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.4.0
