### 2023-07-07 #5627 @Midnighter
FYI, the license is Apache-2.0. It's just not present on the paper branch. Should it be there?
