### 2023-07-08 #5627 @Kevin-Mattheus-Moerman
@Midnighter @jfy133 below are some minor comments on the paper:

- [ ] You may shorten your affiliations to use the simple format `Department, University, City, Country`, so you may remove street names and postal codes.
- [ ] Words like `optimized` and `visualization` are American English whereas words like `characterisation`, `analyse`, `standardise`, `standardised` and `standardising` are British English. Please choose one and use consistently throughout the paper.
