### 2023-07-08 #5627 @jfy133
@Kevin-Mattheus-Moerman - I've addressed both issues in the PR above. I could only find those two American spellings (also after spell check with British dictionary), were there anymore that you noticed? The one exception is the spelling of a project in the funding section, but the official name is with the American spelling of Paleo so I assume this is OK.
