### 2023-07-11 #5627 @Kevin-Mattheus-Moerman
The work submitted here has been reviewed/accepted by `pyOpenSci`, see here:  https://github.com/pyOpenSci/software-submission/issues/84
