### 2023-07-11 #5627 @Kevin-Mattheus-Moerman
@Midnighter @jfy133 please check the above checklist :point_up: and your required steps. In particular for the archive you need to manually edit the author set and order, and the title, to match those for paper. Also ensure the license listed is your software license. Add ORCIDs to the archive too.
