### 2023-07-11 #5627 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
