+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - (2023-07-13)
+
 ### Fixed
 
 -   Forced reading the `duplicates` and `coverage` columns of KrakenUniq profiles as
